requirements.txt
Django==3.0.5
django-heroku
gunicorn
pytube3
opencv-python-headless
ffmpeg-python
requests
